2022 day03 improve
use the ascii value of the char to compute the value. take value of char
and subtract 'a' or 'A' depending on if it's upper or lower and add 1
or 27 depending on if it's high or low side. this removes the need
for a hand build collection of chars to index.

change collection grouping into an iterator rather than using a for
loop which simplifies the code and likely avoids allocations.