2022 day06 improve code
This change collapses the code down into a single set of functions which take
arguments for the width of distinct characters needed to match for the
validation.

The other change is to simplify the check for the validation using the
itertools all_unique method to validate the contents are all unique per
directions. My assumption is that this still could be improved with other
techniques, but I'm happy to have simplified the code to this point.