2022 day07 ideas
instead of the approach where we loop and keep state, lets try streaming
commands. since a command is indicated by a '$' then if we treat input
as a big string splitting on '$' we will be able to process each command
directly as it's encountered. also this creates an empty entry at the
start of the stream we need to skip over.

after this we can group each encountered set of data as the command and
it's possible output. we can split this on line separator giving us the
command in the first entry and the output if any in the remaining ones.